@@ -17,7 +17,7 @@
     }
   ],
   "require": {
-    "php": ">=5.5.9"
+    "php": ">=5.6.0"
   },
   "require-dev": {
     "phpunit/phpunit": "5.5.*"
