@@ -1,13 +1,5 @@
 <?php
-// +----------------------------------------------------------------------
-// | ThinkPHP [ WE CAN DO IT JUST THINK IT ]
-// +----------------------------------------------------------------------
-// | Copyright (c) 2009 http://thinkphp.cn All rights reserved.
-// +----------------------------------------------------------------------
-// | Licensed ( http://www.apache.org/licenses/LICENSE-2.0 )
-// +----------------------------------------------------------------------
-// | Author: liu21st <liu21st@gmail.com>
-// +----------------------------------------------------------------------
+
 namespace Think\Crypt\Driver;
 
 /**
@@ -18,20 +10,20 @@ class Think
 
     /**
      * 加密字符串
-     * @param string $str 字符串
+     * @param string $data 字符串
      * @param string $key 加密key
-     * @param integer $expire 有效期（秒）
+     * @param int $expire 有效期（秒）
      * @return string
      */
     public static function encrypt($data, $key, $expire = 0)
     {
         $expire = sprintf('%010d', $expire ? $expire + time() : 0);
-        $key    = md5($key);
-        $data   = base64_encode($expire . $data);
-        $x      = 0;
-        $len    = strlen($data);
-        $l      = strlen($key);
-        $char   = $str   = '';
+        $key = md5($key);
+        $data = base64_encode($expire.$data);
+        $x = 0;
+        $len = strlen($data);
+        $l = strlen($key);
+        $char = $str = '';
 
         for ($i = 0; $i < $len; $i++) {
             if ($x == $l) {
@@ -45,28 +37,29 @@ class Think
         for ($i = 0; $i < $len; $i++) {
             $str .= chr(ord(substr($data, $i, 1)) + (ord(substr($char, $i, 1))) % 256);
         }
-        return str_replace(array('+', '/', '='), array('-', '_', ''), base64_encode($str));
+
+        return str_replace(['+', '/', '='], ['-', '_', ''], base64_encode($str));
     }
 
     /**
      * 解密字符串
-     * @param string $str 字符串
+     * @param string $data 字符串
      * @param string $key 加密key
      * @return string
      */
     public static function decrypt($data, $key)
     {
-        $key  = md5($key);
-        $data = str_replace(array('-', '_'), array('+', '/'), $data);
+        $key = md5($key);
+        $data = str_replace(['-', '_'], ['+', '/'], $data);
         $mod4 = strlen($data) % 4;
         if ($mod4) {
             $data .= substr('====', $mod4);
         }
         $data = base64_decode($data);
 
-        $x    = 0;
-        $len  = strlen($data);
-        $l    = strlen($key);
+        $x = 0;
+        $len = strlen($data);
+        $l = strlen($key);
         $char = $str = '';
 
         for ($i = 0; $i < $len; $i++) {
@@ -85,12 +78,13 @@ class Think
                 $str .= chr(ord(substr($data, $i, 1)) - ord(substr($char, $i, 1)));
             }
         }
-        $data   = base64_decode($str);
+        $data = base64_decode($str);
         $expire = substr($data, 0, 10);
         if ($expire > 0 && $expire < time()) {
             return '';
         }
         $data = substr($data, 10);
+
         return $data;
     }
 }
