@@ -20,7 +20,7 @@ class File extends Storage
 
     /**
      * 读取文件内容
-     * @access public
+     
      * @internal
      *
      * @param string $filename 文件名
@@ -34,7 +34,7 @@ class File extends Storage
 
     /**
      * 文件写入
-     * @access public
+     
      *
      * @param string $filename 文件名
      * @param string $content 文件内容
@@ -58,7 +58,7 @@ class File extends Storage
 
     /**
      * 文件追加写入
-     * @access public
+     
      *
      * @param string $filename 文件名
      * @param string $content 追加的文件内容
@@ -76,7 +76,7 @@ class File extends Storage
 
     /**
      * 加载文件
-     * @access public
+     
      *
      * @param string $_filename 文件名
      * @param array $vars 传入变量
@@ -94,7 +94,7 @@ class File extends Storage
 
     /**
      * 文件是否存在
-     * @access public
+     
      *
      * @param string $filename 文件名
      *
@@ -107,7 +107,7 @@ class File extends Storage
 
     /**
      * 文件删除
-     * @access public
+     
      *
      * @param string $filename 文件名
      *
@@ -122,7 +122,7 @@ class File extends Storage
 
     /**
      * 读取文件信息
-     * @access public
+     
      *
      * @param string $filename 文件名
      * @param string $name 信息名 mtime或者content
