@@ -45,7 +45,7 @@ final class Cx extends TagLib
 
     /**
      * php标签解析
-     * @access public
+
      *
      * @param array $tag 标签属性
      * @param string $content 标签内容
@@ -64,7 +64,7 @@ final class Cx extends TagLib
      * {user.username}
      * {user.email}
      * </volist>
-     * @access public
+
      *
      * @param array $tag 标签属性
      * @param string $content 标签内容
@@ -107,7 +107,7 @@ final class Cx extends TagLib
 
     /**
      * foreach标签解析 循环输出数据集
-     * @access public
+
      *
      * @param array $tag 标签属性
      * @param string $content 标签内容
@@ -139,7 +139,7 @@ final class Cx extends TagLib
      * <else />
      * </if>
      * 表达式支持 eq neq gt egt lt elt == > >= < <= or and || &&
-     * @access public
+
      *
      * @param array $tag 标签属性
      * @param string $content 标签内容
@@ -156,7 +156,7 @@ final class Cx extends TagLib
     /**
      * else标签解析
      * 格式：见if标签
-     * @access public
+
      *
      * @param array $tag 标签属性
      * @param string $content 标签内容
@@ -172,7 +172,7 @@ final class Cx extends TagLib
 
     /**
      * else标签解析
-     * @access public
+
      *
      * @param array $tag 标签属性
      *
@@ -191,7 +191,7 @@ final class Cx extends TagLib
      * <case value="2" >2</case>
      * <default />other
      * </switch>
-     * @access public
+
      *
      * @param array $tag 标签属性
      * @param string $content 标签内容
@@ -213,7 +213,7 @@ final class Cx extends TagLib
 
     /**
      * case标签解析 需要配合switch才有效
-     * @access public
+
      *
      * @param array $tag 标签属性
      * @param string $content 标签内容
@@ -252,7 +252,7 @@ final class Cx extends TagLib
     /**
      * default标签解析 需要配合switch才有效
      * 使用： <default />some value
-     * @access public
+
      *
      * @param array $tag 标签属性
      *
@@ -267,7 +267,7 @@ final class Cx extends TagLib
      * compare标签解析
      * 用于值的比较 支持 eq neq gt lt egt elt heq nheq 默认是eq
      * 格式： <compare name="" type="eq" value="" >content</compare>
-     * @access public
+
      *
      * @param array $tag 标签属性
      * @param string $content 标签内容
@@ -413,7 +413,7 @@ final class Cx extends TagLib
      * 如果某个变量存在于某个范围 则输出内容 type= in 表示在范围内 否则表示在范围外
      * 格式： <range name="var|function"  value="val" type='in|notin' >content</range>
      * example: <range name="a"  value="1,2,3" type='in' >content</range>
-     * @access public
+
      *
      * @param array $tag 标签属性
      * @param string $content 标签内容
@@ -509,7 +509,7 @@ final class Cx extends TagLib
      * present标签解析
      * 如果某个变量已经设置 则输出内容
      * 格式： <present name="" >content</present>
-     * @access public
+
      *
      * @param array $tag 标签属性
      * @param string $content 标签内容
@@ -528,7 +528,7 @@ final class Cx extends TagLib
      * notpresent标签解析
      * 如果某个变量没有设置，则输出内容
      * 格式： <notpresent name="" >content</notpresent>
-     * @access public
+
      *
      * @param array $tag 标签属性
      * @param string $content 标签内容
@@ -547,7 +547,7 @@ final class Cx extends TagLib
      * empty标签解析
      * 如果某个变量为empty则输出内容
      * 格式： <empty name="" >content</empty>
-     * @access public
+
      *
      * @param array $tag 标签属性
      * @param string $content 标签内容
@@ -614,7 +614,7 @@ final class Cx extends TagLib
      * 在模板中给某个变量赋值 支持变量赋值
      * 格式： <assign name="" value="" />
      * @deprecated
-     * @access public
+
      *
      * @param array $tag 标签属性
      * @param string $content 标签内容
@@ -637,7 +637,7 @@ final class Cx extends TagLib
      * define标签解析
      * 在模板中定义常量 支持变量赋值
      * 格式： <define name="" value="" />
-     * @access public
+
      *
      * @param array $tag 标签属性
      * @param string $content 标签内容
@@ -659,7 +659,7 @@ final class Cx extends TagLib
     /**
      * for标签解析
      * 格式： <for start="" end="" comparison="" step="" name="" />
-     * @access public
+
      *
      * @param array $tag 标签属性
      * @param string $content 标签内容
