@@ -74,10 +74,8 @@ final class Verify
     private $_color = null;
 
     /**
-     * 架构方法 设置参数
-     * @access public
-     *
-     * @param  array $config 配置参数
+     * Verify constructor.
+     * @param array $config 配置参数
      */
     public function __construct(array $config = [])
     {
@@ -86,7 +84,6 @@ final class Verify
 
     /**
      * 使用 $this->name 获取配置
-     * @access public
      *
      * @param  string $name 配置名称
      *
@@ -99,7 +96,6 @@ final class Verify
 
     /**
      * 设置验证码配置
-     * @access public
      *
      * @param  string $name 配置名称
      * @param  string $value 配置值
@@ -115,7 +111,6 @@ final class Verify
 
     /**
      * 检查配置
-     * @access public
      *
      * @param  string $name 配置名称
      *
@@ -128,7 +123,6 @@ final class Verify
 
     /**
      * 验证验证码是否正确
-     * @access public
      *
      * @param string $code 用户验证码
      * @param string $id 验证码标识
@@ -162,7 +156,6 @@ final class Verify
     /**
      * 输出验证码并把验证码的值保存的session中
      * 验证码保存到session的格式为： array('verify_code' => '验证码值', 'verify_time' => '验证码创建时间');
-     * @access public
      *
      * @param string $id 要生成验证码的标识
      *
